TOML 1.0.0: make arrays heterogeneous
This change breaks the public API. This is necessary, as in TOML 1.0.0
and unlike for previous format versions, arrays can hold values of
heterogeneous types (for instance: [1, "foo", True]).

This subtle change breaks the semantics of the API we had so far, which
assumed that arrays could contain only homogeneous values (only
integers, or only booleans, or ...). In order to adapt to the new
semantics, this commit removes the constructor/accessors with this
assumption.